Refactor errors and imports
