+++ packages/probing-eservice-operations/src/services/db/eserviceQuery.ts
@@ -13,7 +13,7 @@ import {
   ReadModelService,
   EServiceQueryFilters,
   EServiceProducersQueryFilters,
-} from "./readModelService.js";
+} from "./dbService.js";
 
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export function eserviceQueryBuilder(readModelService: ReadModelService) {
